moved internal extension methods into the Internal namespace
