+++ ObjectDumper/Internal/ObjectDumperCSharp.cs
@@ -4,8 +4,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-using ObjectDumping.Extensions;
-
 namespace ObjectDumping.Internal
 {
     /// <summary>
