+++ ObjectDumper/Internal/TypeExtensions.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Reflection;
 using System.Text;
 
-namespace ObjectDumping.Extensions
+namespace ObjectDumping.Internal
 {
     internal static class TypeExtensions
     {
